requirements.txt
streamlit==1.33.0
langchain==0.2.12
langchain-community==0.2.12
langchain-text-splitters==0.2.3
langchain-openai==0.1.22
openai==1.40.6
sentence-transformers==2.7.0
faiss-cpu==1.7.4
rank-bm25==0.2.2
pypdf==4.2.0
tiktoken==0.7.0